Enforce that named_tensor_meta_ is non-null only if there is a non-wildcard name
Today, there are two equivalent representations: named_tensor_meta_ is
null, or named_tensor_meta_ is non-null but all of the dimension names
are wildcard.  Let's reduce the opportunity for behavior divergence by
making the second representation illegal.

This will make it easier for me to add a dispatch key for named
tensor as I can rely on setters to always go through TensorImpl to
maintain invariants on DispatchKey.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 621506ee179a71b44e77f577fb8b7b970f0e8a8a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/38725