Simplify precision-specification in tests.
Now that assertEquals considers dtypes in determining tolerance, most
tests don't need explicitly set precision.

Those that do are a few half precision tests on cuda. In this PR, those
are broken out to be handled explicitly, though we may also want to
consider further loosening the tolerance on half-precision.

ghstack-source-id: 554f6db560ed1e1ff6045f90ec335e442b02e0a6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37181